python_flask_api/requirements.txt
Flask
Flask-MySQLdb    